+++ src/core_systems/continuous_time_ode.jl
@@ -136,7 +136,7 @@ for f in (:initial_state, :current_parameters, :dynamic_rule,
 end
 
 SciMLBase.isinplace(::CoupledODEs{IIP}) where {IIP} = IIP
-set_state!(ds::CoupledODEs, u::AbstractArray) = set_state!(ds.integ, u)
+set_state!(ds::CoupledODEs, u::AbstractArray) = (set_state!(ds.integ, u); ds)
 
 # so that `ds` is printed
 SciMLBase.step!(ds::CoupledODEs, args...) = (step!(ds.integ, args...); ds)
@@ -173,7 +173,7 @@ end
 function set_state!(integ::DEIntegrator, u)
     if integ.u isa Array{<:Real}
         integ.u .= u
-    elseif integ.u isa Union{SVector, SMatrix}
+    elseif integ.u isa StateSpaceSets.StaticArraysCore.SArray{<:Real}
         integ.u = u
     else
         integ.u = recursivecopy(u)