+++ src/core_systems/discrete_time_map.jl
@@ -73,7 +73,7 @@ isdeterministic(::DIM) = true
 function set_state!(ds::DeterministicIteratedMap{IIP}, u) where {IIP}
     ds.u = recursivecopy(u)
     ds.dummy = recursivecopy(u)
-    return
+    return ds
 end
 
 ###########################################################################################